move other use cases to dedicated package
