+++ feature/composer/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feature/composer/usecase/PrepareForPreviewUseCase.kt
@@ -1,8 +1,8 @@
-package com.livefast.eattrash.raccoonforfriendica.feature.composer.utils
+package com.livefast.eattrash.raccoonforfriendica.feature.composer.usecase
 
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.data.MarkupMode
 
-interface StripMarkupUseCase {
+interface PrepareForPreviewUseCase {
     operator fun invoke(
         text: String,
         mode: MarkupMode,
